Store slurm job-id in config.
This is needed, as we currently have only very obscure link between slurm job-id (which is used for slurm_logs stdout storing for example) and the running job itself (workdir, xid, wid, ...).
Note that on pre-emptions, the job can get a new slurm job-id, and we do overwrite the config.json on each re-run, so this will always contain the LAST job-id, which is the most meaningful thing, since we'd usually want the slurm ID to do job-control things.