Fix unintended recompiles/graph breaks and noisy warnings.
The big unlock of doing this is not actually "control" or speed, but without this, eventually our code simply dies as we get more diverse evaluators and hit the recompile limit, we'll then get uncompiled flex-attention which simply dies of cuda OOM. I was not able to run finevision with multi-resolution evaluators at all.

A lot of the recompiles are simple in hindsight, but there's a few bigger ones to note:

1. Just wrapping a function doesn't make it different compiles, because PyTorch uses the "code object" to determine recompiles. I had long discussions with the devs, and the `clone_function` way, with importantly the `f.__code__.replace()` to get a new "code object" for it, is the way to get independent compiles of a function.

2. We cannot compile the chunked loss with a working recompiles=1 limit. Because our seqlen for the loss is, for example, 4095, so when we chunk it into 8 pieces, we get [512, 512, ..., 511] sized pieces. So, I marked this whole region as not-compiled. (It creates graph-break due to the `.backward()` anyways.)

3. For SimpleFSDP, param-free modules were also being wrapped in the custom type, and all of them in the same-hash custom type, which was an issue. Now simply skip param-free modules. (And maybe we should skip replicate-only modules too, but that's for another time.)

Here is the speed and memory comparison of different variants:

|                | nightly                | stable                 |
|----------------|------------------------|------------------------|
| **head**       | 449.2ms 8260.4MiB      | 394.2ms 8260.4MiB      |
| **pr (nocompile)** | 449.1ms 8629.3MiB      | 447.3ms 8571.7MiB      |
| **pr (compile)**   | 440.1ms 8627.7MiB      | 387.7ms 8570.3MiB      |

And also the change of `chunks` to `chunksz` is a no-op: `387.7ms 8570.3MiB` -> `387.2ms 8570.3MiB`.